Remove usememo and refactor
