+++ template/src/lib/sdk/models/CategoryModel.js
@@ -1,4 +1,4 @@
-import { executeSQL, getFilterCondition, getConditionFromViewPort } from '..';
+import { executeSQL, filtersToSQL, viewportToSQL } from '..';
 
 export const getCategories = async (props) => {
   const { data, credentials, column, operation, filters, viewport } = props;
@@ -10,8 +10,7 @@ export const getCategories = async (props) => {
   }
 
   let query =
-    (viewport &&
-      `SELECT * FROM (${data})  as q WHERE ${getConditionFromViewPort(viewport)}`) ||
+    (viewport && `SELECT * FROM (${data})  as q WHERE ${viewportToSQL(viewport)}`) ||
     data;
 
   query = `WITH all_categories as (
@@ -22,7 +21,7 @@ export const getCategories = async (props) => {
   categories as (
     SELECT ${column} as category, ${operation}(${operationColumn}) as value
       FROM (${query}) as q
-    ${getFilterCondition(filters)}
+    ${filtersToSQL(filters)}
     GROUP BY category
   )
   SELECT a.category, b.value
