+++ template/src/lib/sdk/models/index.js
@@ -4,7 +4,7 @@ export { getValue } from './FormulaModel';
 export { getHistogram } from './HistogramModel';
 export {
   FilterTypes,
-  getFilterCondition,
-  getConditionFromViewPort,
-  getFilteredQuery,
-} from './FilterConditionBuilder';
+  filtersToSQL,
+  viewportToSQL,
+  buildQuery,
+} from './FilterQueryBuilder';
