+++ template/src/lib/sdk/widgets/CategoryWidget.js
@@ -1,9 +1,15 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { selectSourceById, addFilter, removeFilter } from 'config/cartoSlice';
 import { FilterTypes, getCategories } from 'lib/sdk';
 import { WrapperWidgetUI, CategoryWidgetUI } from '@carto/react-airship-ui';
-import { getApplicableFilters } from '../models/FilterConditionBuilder';
+import { getApplicableFilters } from '../models/FilterQueryBuilder';
+
+function getSelectCategories(column, filters) {
+  return filters && filters[column] && filters[column][FilterTypes.IN]
+    ? filters[column][FilterTypes.IN].values
+    : [];
+}
 
 export default function CategoryWidget(props) {
   const { column } = props;
@@ -17,13 +23,7 @@ export default function CategoryWidget(props) {
     (state) => selectSourceById(state, props['data-source']) || {}
   );
   const { data, credentials } = source;
-  const { filters: _filters = {} } = source,
-    { [column]: _column = {} } = _filters,
-    { [FilterTypes.IN]: { values: selectedCategories = [] } = {} } = _column;
-
-  const filters = useMemo(() => {
-    return getApplicableFilters(_filters, props.id);
-  }, [_filters, props.id]);
+  const selectedCategories = getSelectCategories(column, source.filters);
 
   useEffect(() => {
     if (
@@ -31,6 +31,7 @@ export default function CategoryWidget(props) {
       credentials &&
       (!props['viewport-filter'] || (props['viewport-filter'] && viewport))
     ) {
+      const filters = getApplicableFilters(source.filters, props.id);
       setLoading(true);
       getCategories({ ...props, data, filters, credentials, viewport }).then((data) => {
         setCategoryData(data);
@@ -39,7 +40,7 @@ export default function CategoryWidget(props) {
     } else {
       setCategoryData([]);
     }
-  }, [credentials, data, filters, viewport, props]);
+  }, [credentials, data, source.filters, viewport, props]);
 
   const handleSelectedCategoriesChange = (categories) => {
     if (categories && categories.length) {
